moved scan process to internal pakage
